Move more BaseEntity logic to EntityHelper, move entity interfaces to entity decorator
